+++ .storybook/main.cjs
@@ -1,23 +1,9 @@
-const storyIndexers = require("./indexer");
-const path = require("path");
-const { mergeConfig } = require("vite");
-
 module.exports = {
-  storyIndexers,
-  devServer: {
-    stats: "errors-only",
-  },
   staticDirs: ["./public"],
 
   stories: () => [
     "../@navikt/**/*.stories.@(js|jsx|ts|tsx|mdx)",
     "./*.stories.mdx",
-    ...(process.env.STORYBOOK_STORIES === "all"
-      ? [
-          "../aksel.nav.no/website/components/**/*.stories.tsx",
-          "../aksel.nav.no/website/pages/eksempler/**/*.tsx",
-        ]
-      : []),
   ],
   addons: [
     "@storybook/addon-a11y",
@@ -55,30 +41,5 @@ module.exports = {
     check: false,
     checkOptions: {},
     reactDocgen: false,
-  },
-  async viteFinal(config) {
-    return mergeConfig(config, {
-      define: {
-        "process.env": {},
-      },
-      resolve: {
-        alias: [
-          {
-            find: "components",
-            replacement: path.resolve(
-              __dirname,
-              "../aksel.nav.no/website/components"
-            ),
-          },
-          {
-            find: "@utils",
-            replacement: path.resolve(
-              __dirname,
-              "../aksel.nav.no/website/components/website-modules/utils/index.ts"
-            ),
-          },
-        ],
-      },
-    });
   },
 };
