+++ aksel.nav.no/website/.storybook/preview.ts
@@ -1,5 +1,8 @@
 import type { Preview } from "@storybook/react";
 
+/* TODO: update import to your tailwind styles file. If you're using Angular, inject this through your angular.json config instead */
+import "../dist/tw.css";
+
 const preview: Preview = {
   parameters: {
     actions: { argTypesRegex: "^on[A-Z].*" },
