+++ aksel.nav.no/website/package.json
@@ -25,6 +25,7 @@
     "playwright": "playwright test",
     "test:watch": "jest --watch",
     "backup": "ts-node scripts/backup.js",
+    "storybook:aksel": "yarn concurrently \"yarn watch:tw\" \"storybook dev -p 6006\"",
     "watch:tw": "yarn tailwindcss -i ./styles/index.css -o ./dist/tw.css --watch"
   },
   "dependencies": {
@@ -88,6 +89,7 @@
     "@types/react-dom": "^18.0.0",
     "autoprefixer": "^10.4.0",
     "babel-loader": "^8.2.0",
+    "concurrently": "7.2.1",
     "cross-env": "^7.0.0",
     "dotenv": "^10.0.0",
     "jest": "^29.0.0",
